feat: Locale set-up by URL and by User Model
The Locale is set-up by an /xx/ prefix on URL or by the locale of
the User Model. If none is available -or valid- it is set to a
default. Also a config/locale.php was added to specify the
supported locales and the languages associated with each country.
All the localized Routes go throgh a /xx/ prefix in routes/web.php.

Added code:

 - Middleware: SetDefaultLocaleForURLs, SetLocale

 - Service Porvider: LocaleServiceProvider

 - Class: App\Logos\Class

 - Config: config/locale.php

 - Migration & Model: Added counntry and locale fields  to the
   users table and the User Model.